+++ src/util/toError.ts
@@ -1,7 +1,13 @@
-export function toError<TErrorExtensions = { status?: number }>(
-  err: unknown
-): Error & TErrorExtensions {
-  return (
-    err instanceof Error ? err : new Error(`Unknown error: ${err}`)
-  ) as Error & TErrorExtensions;
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+export function toError(err: any): Error {
+  if (
+    err &&
+    typeof err === "object" &&
+    typeof err.message === "string" &&
+    typeof err.stack === "string"
+  ) {
+    return err;
+  }
+
+  return new Error(`Unknown error: ${JSON.stringify(err)}`);
 }